Extract common methods from ParseAddProperty and ParseAddClient into new superclass ParseAdd (ParserAdd Parent Class is Parser)
