+++ src/main/java/seedu/duke/parsermanager/ParseAddProperty.java
@@ -2,7 +2,6 @@ package seedu.duke.parsermanager;
 
 import seedu.duke.command.Command;
 import seedu.duke.command.CommandAddProperty;
-import seedu.duke.exception.EmptyDescriptionException;
 import seedu.duke.exception.EmptyDetailException;
 import seedu.duke.exception.IncorrectFlagOrderException;
 import seedu.duke.exception.InvalidPriceFormatException;
@@ -10,8 +9,6 @@ import seedu.duke.exception.InvalidSingaporeAddressException;
 import seedu.duke.exception.MissingFlagException;
 
 import java.util.ArrayList;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import static seedu.duke.CommandStructure.ADD_PROPERTY_FLAGS;
 import static seedu.duke.Messages.EXCEPTION;
@@ -19,14 +16,13 @@ import static seedu.duke.Messages.MESSAGE_ADD_PROPERTY_WRONG_FORMAT;
 import static seedu.duke.Messages.MESSAGE_INVALID_SINGAPORE_ADDRESS;
 import static seedu.duke.Messages.MESSAGE_INVALID_PRICE_FORMAT;
 
-public class ParseAddProperty extends Parser {
+public class ParseAddProperty extends ParseAdd {
     private final String commandDescription;
 
     private static final int PROPERTY_FLAG_SIZE = 4;
     private static final int PROPERTY_ADDRESS_INDEX = 1;
     private static final int PROPERTY_PRICE_INDEX = 2;
 
-    private static final int MISSING_FLAG_VALUE = -1;
     private static final int FLAG_JUMPER_VALUE = 2;
     private static final int UNIT_VALUE = 1;
 
@@ -67,8 +63,8 @@ public class ParseAddProperty extends Parser {
         this.commandDescription = addCommandDescription;
     }
 
-    public Command parseCommand() throws EmptyDescriptionException, InvalidSingaporeAddressException,
-            MissingFlagException, IncorrectFlagOrderException, InvalidPriceFormatException, EmptyDetailException {
+    public Command parseCommand() throws  EmptyDetailException, MissingFlagException, IncorrectFlagOrderException,
+            InvalidSingaporeAddressException, InvalidPriceFormatException {
         try {
             checkForEmptyDetails(commandDescription);
             ArrayList<String> propertyDetails = processCommandAddPropertyDetails(commandDescription);
@@ -87,57 +83,25 @@ public class ParseAddProperty extends Parser {
         }
     }
 
-    private void checkForEmptyDetails(String commandDetail) throws EmptyDetailException {
-        boolean isEmptyDetail = isEmptyString(commandDetail);
-        if (isEmptyDetail) {
-            throw new EmptyDetailException(EXCEPTION);
-        }
-    }
-
-    private boolean isEmptyString(String commandDetail) {
-        return commandDetail.trim().isEmpty();
-    }
-
     private ArrayList<String> processCommandAddPropertyDetails(String rawCommandDetail)
             throws MissingFlagException, IncorrectFlagOrderException {
-        String[] flags = ADD_PROPERTY_FLAGS;
-        int[] flagIndexPositions = getFlagIndexPositions(rawCommandDetail, flags);
+        int[] flagIndexPositions = getFlagIndexPositions(rawCommandDetail, ADD_PROPERTY_FLAGS);
         checkForMissingPropertyFlags(flagIndexPositions);
         checkPropertyFlagsOrder(flagIndexPositions);
         return extractPropertyDetails(rawCommandDetail, flagIndexPositions);
     }
 
-    private int[] getFlagIndexPositions(String commandDetail, String[] flags) {
-        int[] flagIndexPositions = new int[flags.length];
-
-        for (int flagIndex = 0; flagIndex < flags.length; flagIndex++) {
-            flagIndexPositions[flagIndex] = commandDetail.indexOf(flags[flagIndex]);
-        }
-        return flagIndexPositions;
-    }
-
     private void checkForMissingPropertyFlags(int[] flagIndexPositions) throws MissingFlagException {
         for (int flagIndex : flagIndexPositions) {
-            if (!isFlagPresent(flagIndex)) {
+            if (!checkForFlagPresence(flagIndex)) {
                 throw new MissingFlagException(EXCEPTION);
             }
         }
     }
 
-    private boolean isFlagPresent(int flagIndexPosition) {
-        return (flagIndexPosition != MISSING_FLAG_VALUE);
-    }
-
     private void checkPropertyFlagsOrder(int[] flagIndexPositions) throws IncorrectFlagOrderException {
         for (int flagIndex = 0; flagIndex < PROPERTY_FLAG_SIZE - UNIT_VALUE; flagIndex++) {
             checkForCorrectFlagOrder(flagIndexPositions[flagIndex], flagIndexPositions[flagIndex + UNIT_VALUE]);
-        }
-    }
-
-    private void checkForCorrectFlagOrder(int flagPosition, int nextFlagPosition) throws IncorrectFlagOrderException {
-        boolean hasCorrectOrder = (flagPosition < nextFlagPosition);
-        if (!hasCorrectOrder) {
-            throw new IncorrectFlagOrderException(EXCEPTION);
         }
     }
 
@@ -155,14 +119,6 @@ public class ParseAddProperty extends Parser {
             }
         }
         return extractedPropertyDetails;
-    }
-
-    private static String extractDetail(String rawDetail, int beginIndex) {
-        return rawDetail.substring(beginIndex).trim();
-    }
-
-    private static String extractDetail(String rawDetail, int beginIndex, int endIndex) {
-        return rawDetail.substring(beginIndex, endIndex).trim();
     }
 
     private void validatePropertyDetails(ArrayList<String> propertyDetails) throws EmptyDetailException,
@@ -211,11 +167,5 @@ public class ParseAddProperty extends Parser {
         if (!hasValidPriceNumberFormat) {
             throw new InvalidPriceFormatException(EXCEPTION);
         }
-    }
-
-    private boolean checkForDetailFormat(String regex, String detail) {
-        Pattern pattern = Pattern.compile(regex);
-        Matcher matcher = pattern.matcher(detail);
-        return matcher.matches();
     }
 }
